fix: lint & prettier 적용
web/app/page.tsx 코드에 lint & prettier 적용.
기존 apps/user 디렉토리 삭제
